Refactor widget
